"deleteUserById function works"

+++ simpleGETproject/simpleget/src/main/java/com/alexdev/simpleget/repositorys/BookRepository.java
@@ -8,7 +8,7 @@ import com.alexdev.simpleget.entity.Book;
 
 public interface BookRepository extends CrudRepository<Book, Long> {
     public ArrayList<Book> findAll();
-    public Book findById(long id);
+    public Book findById(Integer id);
 
     public Book save(Book newBook);
